% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecks.R
\name{CheckArrayDim}
\alias{CheckArrayDim}
\title{Check the dimensions of a power array}
\usage{
CheckArrayDim(x, required_dim = NULL, condition = "warning", sliced = FALSE)
}
\arguments{
\item{x}{A power array}

\item{required_dim}{The number of dimensions required. If NULL just returns
the number of dimensions.}

\item{condition}{If required dimensions are not correct should a \code{warning} or
\code{error} be produced. Ignored if required dimensions is NULL.}

\item{sliced}{For the condition text should the user be informed the array
is sliced. Ignored if required dimensions is NULL.}
}
\value{
An integer for the number of dimensions
}
\description{
Check the dimensions of a power array
}
