% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecks.R
\name{EnsureSingleFunOut}
\alias{EnsureSingleFunOut}
\title{Ensure a power_array has only a single fun_out value}
\usage{
EnsureSingleFunOut(x, condition = "warning", sliced = FALSE)
}
\arguments{
\item{x}{A power array}

\item{condition}{If array is not summarised should a \code{warning} or
\code{error} be produced. Ignored if required dimensions is NULL.}
}
\value{
A power array with only a single fun_out value
}
\description{
If multiple fun_out values are present in the input take the first one.
}
