% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecks.R
\name{EnsureSummarized}
\alias{EnsureSummarized}
\title{Ensure a power_array is summarized}
\usage{
EnsureSummarized(x, summary_function = NULL, condition = "warning")
}
\arguments{
\item{x}{A power array}

\item{summary_function}{What function should be used to summarize the array,
if it is not already summarized.}

\item{condition}{If array is not summarised should a \code{warning} or
\code{error} be produced. Ignored if required dimensions is NULL.}
}
\value{
A summarised power_array
}
\description{
Ensure a power_array is summarized
}
