% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpowergrid.R
\name{GetPowergrid}
\alias{GetPowergrid}
\title{Get core element from \code{powCal} or \code{powEx} object}
\usage{
GetPowergrid(x)
}
\arguments{
\item{x}{Output object from \code{sse::powCal} or \code{sse::powEx}}
}
\value{
array with appropriately named dimensions
}
\description{
Get from objects resulting from \code{sse::powCal} (class
power) or \code{sse::powEx} (class powCalc) the core element and supply
with appropriate dimnames.
}
\author{
Gilles Dutilh
}
