% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergrid.R
\name{print.power_array}
\alias{print.power_array}
\title{print}
\usage{
\method{print}{power_array}(x, ...)
}
\arguments{
\item{x}{object of class power_array}

\item{...}{passed on to \code{cat}}
}
\description{
Method for printing objects of class power_array.
##'
Prints a power_array as a default array with a short summary about
its contents.
}
\seealso{
\code{\link{PowerGrid}}
}
\author{
Gilles Dutilh
}
