% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{print.power_example}
\alias{print.power_example}
\title{Print example}
\usage{
\method{print}{power_example}(x, ...)
}
\arguments{
\item{x}{object of class \code{power_example}}

\item{...}{passed on to \code{cat}}
}
\value{
nothing
}
\description{
Print method for class \code{power_example}.
}
\details{
Print short informative output for object of class \code{power_example}.
}
\author{
Gilles Dutilh
}
