% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{PrintWrap}
\alias{PrintWrap}
\title{Wrap text in printed output.}
\usage{
PrintWrap(x, width = 48, ...)
}
\arguments{
\item{x}{A single character string}

\item{width}{The desired width}

\item{...}{Further arguments passed to 'strwrap()'}
}
\value{
a character string with returns placed such to achieve the desired
width
}
\description{
Wrap text in printed output. This function is created for
convenience and also so that the width of wrapping can be set by simply
changing the default here. Note that this give a leading white space to
each resulting line. The is done to account for the fact that wrapping
deletes the whites paces, and it actually looks good.
}
\author{
Gilles Dutilh
}
