% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergrid.R
\name{SummarizeIterations}
\alias{SummarizeIterations}
\title{Summary of object that has individual iterations saved.}
\usage{
SummarizeIterations(x, summary_function, ...)
}
\arguments{
\item{x}{Object of class \code{power_array}}

\item{summary_function}{function to apply across iterations}

\item{...}{Further arguments passed to 'summary_function'}
}
\value{
An object of class \code{power_array}, with attributes \code{summarized =
  TRUE}.
}
\description{
Summarizes objects of class \code{power_array} that have individual
iterations saved.
}
\seealso{
\code{\link{PowerGrid}}
}
\author{
Gilles Dutilh
}
