% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{summary.power_example}
\alias{summary.power_example}
\title{Print Example}
\usage{
\method{summary}{power_example}(x)
}
\arguments{
\item{x}{object of class \code{power_example}}
}
\value{
nothing
}
\description{
Summary method for class \code{power_example}.
}
\details{
Print longer informative output for object of class \code{power_example}.
}
\author{
Gilles Dutilh
}
