% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test}
\alias{test}
\title{the name}
\usage{
test(x, bla)
}
\arguments{
\item{x}{describes gilles}

\item{bla}{does nothing}
}
\value{
nothing
}
\description{
This is the decription, apparently
}
\details{
and here are some details.
}
\author{
me myself
}
