% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplot.R
\name{AddExample}
\alias{AddExample}
\title{Add an example to an existing power plot}
\usage{
AddExample(
  x,
  slicer = NULL,
  example,
  target = 0.9,
  minimal_target = TRUE,
  find_min = TRUE,
  method = "step",
  col = 1,
  ...
)
}
\arguments{
\item{x, slicer, example, target, minimal_target, find_min, method}{See help for
\code{PowerPlot}.}

\item{col}{Color or arrow drawn.}

\item{...}{Further arguments are passed to the two calls of function
\code{graphics::arrows} drawing the nicked arrow.}
}
\value{
Nothing
}
\description{
Add an example arrow to an existing power plot created by
PowerPlot.
}
\seealso{
\code{\link{PowerPlot}}, \code{\link{GridPlot}}
}
\author{
Gilles Dutilh
}
