% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergrid.R
\name{print.power_array}
\alias{print.power_array}
\title{print}
\usage{
\method{print}{power_array}(x)
}
\arguments{
\item{x}{object of class power_array}
}
\description{
Method for printing objects of class power_array.
##'
Prints a power_array as a default array with a short summary about
its contents.
}
\author{
Gilles Dutilh
}
