% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ScaleRange}
\alias{ScaleRange}
\title{Scale to desired range}
\usage{
ScaleRange(x, a, b)
}
\arguments{
\item{x}{A vector}

\item{a}{The desired minimum}

\item{b}{The desired maxmum}
}
\value{
a vector with the desired minimum and maximum
}
\description{
Linear scaling of a vector to a vector with a required range.
}
\author{
Gilles Dutilh
}
