% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergrid.R
\name{[.power_array}
\alias{[.power_array}
\title{indexing with [ ] for class \code{power_array}

[ ]: R:\%20}
\usage{
\method{[}{power_array}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{object}

\item{...}{index}

\item{drop}{drop}
}
\value{
An array of class \code{power_grid}
}
\description{
Method for indexing [] of objects of class power_array. The method
makes sure that the resulting array is of class power_array and
keeps and updates the object's attributes. These attributes are
needed for various functions in the powergrid package to work
well.
##'
The indexing functions as normal indexing, but note that drop is
FALSE by default, so that the resulting array has the same
dimensions as the original array. The number of levels at each
dimension may be reduced, however.
##'
}
\author{
Gilles Dutilh
}
