% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergrid.R
\name{summary.power_array}
\alias{summary.power_array}
\title{Summary of power_grid object.}
\usage{
\method{summary}{power_array}(object, ...)
}
\arguments{
\item{object}{array of class power_grid}

\item{...}{passed on to \code{cat}}
}
\description{
Offers a short summary of the power_array object, summarizing the
range of observed values and the grid evaluated across.
##'
See PowerGrid for details
}
\author{
Gilles Dutilh
}
