% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{summary.power_example}
\alias{summary.power_example}
\title{Print contents of an example}
\usage{
\method{summary}{power_example}(object, ...)
}
\arguments{
\item{object}{object of class \code{power_example}}

\item{...}{passed on to \code{data.frame} (which is the thing that is printed)}
}
\value{
nothing
}
\description{
Summary method for class \code{power_example}.
}
\details{
Print longer informative output for object of class \code{power_example}.
}
\author{
Gilles Dutilh
}
